updated to fix issues complained by valgrind in AtmiBrokerEnv
git-svn-id: http://anonsvn.jboss.org/repos/blacktie/trunk@1087 aa7494b7-e75f-0410-880c-f61b6724bf28

+++ atmibroker-admin/src/main/cpp/AtmiBrokerAdmin.cxx
@@ -23,7 +23,7 @@ void ADMIN(TPSVCINFO* svcinfo) {
 	ACE_OS::strncpy(req, svcinfo->data, svcinfo->len);
 
 	char* toReturn = NULL;
-	int   len = 1;
+	long  len = 1;
 	toReturn = tpalloc((char*) "X_OCTET", NULL, len);
 	toReturn[0] = '0';
